fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
PyJWT
python-json-logger
psycopg2-binary
alembic
bcrypt==4.3.0
pytest
pydantic[email]